Refactor populateDb putting inside populateDB whit the other classes related to populate the db
